:sparkles: 优化输入框功能
