--- src/app/chat/VirtualIdol/style.ts
@@ -1,10 +1,0 @@
-import { createStyles } from 'antd-style';
-
-export const useStyles = createStyles(({ css }) => ({
-  container: css`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  `,
-}));
