+++ src/features/ChatInput/Actions/Dance.tsx
@@ -1,0 +1,18 @@
+import { ActionIcon } from '@lobehub/ui';
+import { Music2 } from 'lucide-react';
+
+import { useConfigStore } from '@/store/config';
+
+export default () => {
+  const [openPanel] = useConfigStore((s) => [s.openPanel]);
+
+  return (
+    <ActionIcon
+      icon={Music2}
+      onClick={() => {
+        openPanel('dance');
+      }}
+      title={'歌舞唱见'}
+    />
+  );
+};
