+++ src/features/ChatInput/Actions/Voice.tsx
@@ -1,11 +1,20 @@
 import { ActionIcon } from '@lobehub/ui';
+import { createStyles } from 'antd-style';
 import classNames from 'classnames';
 import { Volume2 } from 'lucide-react';
 
 import { toogleVoice } from '@/services/chat';
 import { useSessionStore } from '@/store/session';
 
-import { useStyles } from './style';
+const useStyles = createStyles(({ token, css }) => ({
+  voice: css`
+    cursor: pointer;
+    transition: color 0.3s;
+  `,
+  voiceOn: css`
+    color: ${token.colorLinkActive};
+  `,
+}));
 
 const VoiceSwitch = () => {
   const { styles } = useStyles();
